requirements.txt
alembic==1.12.0
fastapi==0.111.0
psycopg2-binary==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
